Week-1/1_model_compare/requirements.txt
langchain
langchain_aws
boto3
langchain_ollama
langchain_groq
python-dotenv
langchain_openai
langchain-google-genai